requirements.txt
streamlit==1.21.0
tensorflow==2.16.1  # Latest stable version (requires Python 3.9-3.11)
numpy==1.23.5
joblib==1.3.0
Pillow==9.4.0